`remotion`: Fix bug where React was included in `remotion/no-react` entry point
Also adding test
