+++ packages/core/src/remotion-environment-context.ts
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export type RemotionEnvironment = {
+	isStudio: boolean;
+	isRendering: boolean;
+	isClientSideRendering: boolean;
+	isPlayer: boolean;
+	isReadOnlyStudio: boolean;
+};
+
+export const RemotionEnvironmentContext =
+	React.createContext<RemotionEnvironment | null>(null);
